+++ editor/vox/vox_import_funcs.cpp
@@ -30,7 +30,7 @@ Ref<Mesh> build_mesh(const VoxelBufferInternal &voxels, VoxelMesher &mesher,
 		std::vector<unsigned int> &surface_index_to_material, Ref<Image> &out_atlas, float p_scale, Vector3 p_offset) {
 	//
 	VoxelMesher::Output output;
-	VoxelMesher::Input input = { voxels, 0 };
+	VoxelMesher::Input input = { voxels, nullptr, nullptr, Vector3i(), 0, false };
 	mesher.build(output, input);
 
 	if (output.surfaces.size() == 0) {