+++ server/mesh_block_task.cpp
@@ -176,7 +176,8 @@ void MeshBlockTask::run(zylann::ThreadedTaskContext ctx) {
 
 	const Vector3i origin_in_voxels = position * (int(data_block_size) << lod);
 
-	const VoxelMesher::Input input = { voxels, meshing_dependency->generator.ptr(), data.get(), origin_in_voxels, lod };
+	const VoxelMesher::Input input = { voxels, meshing_dependency->generator.ptr(), data.get(), origin_in_voxels, lod,
+		collision_hint };
 	mesher->build(_surfaces_output, input);
 
 	_has_run = true;